requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
_libgcc_mutex=0.1
absl-py=0.8.0
astor=0.8.0
ca-certificates=2024
certifi=2020
chardet=3
cudatoolkit=10.0
cycler=0.10
decorator=4.4.0
gast=0.2.2
grpcio=1.23
h5py=2.10
hdf5=1.8
idna=2.8
kiwisolver=1.1
markdown=3.1
networkx=2.3
numpy=1.17
pillow==6.2
pip=10.0
protobuf=3.10
python=3.5
setuptools=50.3
simpleitk=1.2.3
six=1.13
sqlite=3.31
tensorflow-gpu=1.15.0
werkzeug=0.16
xorg-kbproto=1.0
xorg-libice=1.1
xorg-libsm=1.2
xorg-libx11=1.8
xorg-libxau=1.0
xorg-libxdmcp=1.1
xorg-libxext=1.3
xorg-libxrender=0.9
xorg-renderproto=0.11
xorg-xextproto=7.3
xorg-xproto=7.0
xz=5.4
